[issue-5013] [BE][FE] fix: address error filter review
