+++ apps/opik-backend/src/main/java/com/comet/opik/domain/evaluators/TraceFilterEvaluationService.java
@@ -1,17 +1,14 @@
 package com.comet.opik.domain.evaluators;
 
-import com.comet.opik.api.ErrorInfo;
 import com.comet.opik.api.Trace;
 import com.comet.opik.api.filter.Field;
 import com.comet.opik.api.filter.Filter;
-import com.comet.opik.api.filter.Operator;
 import com.comet.opik.api.filter.TraceField;
 import com.comet.opik.api.filter.TraceFilter;
 import jakarta.inject.Inject;
 import jakarta.inject.Singleton;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
 
 import java.util.List;
 
@@ -54,8 +51,7 @@ public class TraceFilterEvaluationService extends FilterEvaluationServiceBase<Tr
     @Override
     public boolean matchesFilter(Filter filter, Trace trace) {
         if (isErrorInfoTextFilter(filter)) {
-            var errorInfoText = extractErrorInfoText(trace.errorInfo());
-            return errorInfoText != null && evaluateOperator(filter.operator(), errorInfoText, filter.value());
+            return matchesErrorInfoTextFilter(filter, trace.errorInfo());
         }
 
         return super.matchesFilter(filter, trace);
@@ -94,41 +90,6 @@ public class TraceFilterEvaluationService extends FilterEvaluationServiceBase<Tr
             case CUSTOM -> extractCustomFieldValue(key, trace);
             default -> {
                 log.warn("Unsupported trace field for filter evaluation: {}", traceField);
-                yield null;
-            }
-        };
-    }
-
-    private boolean isErrorInfoTextFilter(Filter filter) {
-        return filter != null
-                && filter.field() == TraceField.ERROR_INFO
-                && (filter.operator() == Operator.CONTAINS || filter.operator() == Operator.NOT_CONTAINS);
-    }
-
-    private String extractErrorInfoText(ErrorInfo errorInfo) {
-        if (!hasNonEmptyErrorInfo(errorInfo)) {
-            return null;
-        }
-        return extractStringFromJson(errorInfo);
-    }
-
-    private boolean hasNonEmptyErrorInfo(ErrorInfo errorInfo) {
-        return errorInfo != null
-                && (StringUtils.isNotBlank(errorInfo.exceptionType())
-                        || StringUtils.isNotBlank(errorInfo.message())
-                        || StringUtils.isNotBlank(errorInfo.traceback()));
-    }
-
-    private Object extractErrorInfoField(ErrorInfo errorInfo, String key) {
-        if (errorInfo == null || key == null) {
-            return null;
-        }
-        return switch (key.toLowerCase()) {
-            case "exceptiontype", "exception_type" -> errorInfo.exceptionType();
-            case "message" -> errorInfo.message();
-            case "traceback" -> errorInfo.traceback();
-            default -> {
-                log.warn("Unknown ErrorInfo field key: '{}'. Supported keys: exceptionType, message, traceback", key);
                 yield null;
             }
         };
