+++ apps/opik-frontend/tsconfig.json
@@ -37,6 +37,7 @@
   },
   "include": [
     "src",
+    "tests",
     "e2e/**/*.ts",
     "playwright.config.ts"
   ],
